Moves populate to solver
- Links populate to top-level Case to maintain compatibility
